Update to ares v119r03 release.
A working VDP FIFO implementation (write slots only), but probably wildly
inaccurate. Rough timing approximations for the slots.

This build fixes some issues in MD Sonic the Hedgehog (forgot to undo the
background color register bit twiddling when reverting some overly aggressive
VDP-performance optimizations) and also fixes MD Golden Axe II.

I spent a few hours trying to fix Lemmings 2 from locking up immediately. What
happens is I get a fourth Hblank IRQ before it clears a RAM value, whereas
Mednafen only gets three. And that desyncs something and it never gets out of a
loop.

+++ ares/component/processor/arm7tdmi/disassembler.cpp
@@ -30,7 +30,7 @@ auto ARM7TDMI::disassembleInstruction(maybe<n32> pc, maybe<boolean> thumb) -> st
 auto ARM7TDMI::disassembleContext() -> string {
   string output;
   for(u32 n : range(16)) {
-    output.append(_r[n], ":", hex(r(n), 8L), " ");
+    output.append(_r[n], ":", hex((u32)r(n), 8L), " ");
   }
 
   output.append("cpsr:");
